webclient/requirements.txt
Django==5.1.1
psycopg2-binary==2.9.9
simplejson==3.19.3
django-crispy-forms==2.3
crispy-bootstrap4
django-recaptcha==4.0.0
django_registration==3.3
django-filter==24.3
django-tables2==2.6.0
django-tables2-column-shifter==2.2.0
django-compressor==4.5.1
django-libsass==0.9 
django-autocomplete-light==3.11.0
django-widget-tweaks==1.5.0
distlib==0.3.8
django-rosetta==0.10.1
django-auto-logout==0.5.1
reportlab==4.2.4
PyRTF3==0.47.5
pypdf==4.3.1
Pillow==10.4.0
beautifulsoup4==4.12.3
django-extensions==3.2.3
celery==5.4.0
django-add-default-value==0.10.0
selenium==4.25.0
redis==5.1.0
django-celery-beat==2.7.0
django-celery-results==2.5.1
django-prometheus==2.3.1
python-logstash==0.4.8
djangorestframework==3.15.2
djangorestframework-xml==2.0.0
lxml==5.3.0
python-magic==0.4.27
pdf2image==1.17.0
phonenumbers==8.13.46
py7zr==0.22.0
rarfile==4.2
django-cacheops==7.0.2
django-model-utils==5.0.0
pandas==2.2.3
tablib[all]==3.6.1
pre-commit==3.8.0
isort==5.13.2
coverage==7.6.1
django-debug-toolbar==4.4.6
black==24.8.0
django-livereload-server==0.5.1
django-debug-toolbar-template-profiler==2.1.0
psycopg2==2.9.9
django-object-actions==4.3.0
uWSGI==2.0.27
requests==2.32.3
polib==1.2.0
pytz==2024.1
python-dateutil==2.9.0.post0
setuptools==75.1.0
